# apps/rpg-cyborgs/openapi/rpg.json
{
  "openapi": "3.0.1",
  "info": {
      "title": "Rpg API v1",
      "version": "1.0"
  },
  "paths": {
...
      "/api/rpg/{system}/action/args": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepArgs"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepArgs"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepArgs"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepArgs"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_RpgArg[]"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_RpgArg[]"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_RpgArg[]"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/action/autocomplete": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/action/complete": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/action/cost": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/action/outcome": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/action/perform": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActionStepRun"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/action/reset": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityComplete"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Action2"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/activity/create": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityCreate"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityCreate"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityCreate"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ActivityCreate"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Activity2"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Activity2"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Activity2"
                              }
                          }
                      }
                  }
              }
          }
      },
...
  },
  "components": {
      "schemas": {
...
          "Cyborgs.MeleeWeapon": {
              "required": [
                  "damage",
                  "hitBonus"
              ],
              "type": "object",
              "allOf": [
                  {
                      "$ref": "#/components/schemas/RpgEntity"
                  }
              ],
              "properties": {
                  "damage": {
                      "$ref": "#/components/schemas/Values.Dice"
                  },
                  "hitBonus": {
                      "type": "integer",
                      "format": "int32",
                      "readOnly": true
                  }
              },
              "additionalProperties": false
          },
...
          "Reflection.Args.DefaultArg": {
              "type": "object",
              "allOf": [
                  {
                      "$ref": "#/components/schemas/Reflection.Args.RpgArg"
                  }
              ],
              "additionalProperties": false
          },
          "Reflection.Args.DiceArg": {
              "type": "object",
              "allOf": [
                  {
                      "$ref": "#/components/schemas/Reflection.Args.RpgArg"
                  }
              ],
              "additionalProperties": false
          },
          "Reflection.Args.IntegerArg": {
              "type": "object",
              "allOf": [
                  {
                      "$ref": "#/components/schemas/Reflection.Args.RpgArg"
                  }
              ],
              "additionalProperties": false
          },
...
          "Reflection.Args.RpgObjectArg": {
              "type": "object",
              "allOf": [
                  {
                      "$ref": "#/components/schemas/Reflection.Args.RpgArg"
                  }
              ],
              "additionalProperties": false
          },
...
          "RpgGraphState": {
              "required": [
                  "entities",
                  "time"
              ],
              "type": "object",
              "properties": {
                  "entities": {
                      "type": "array",
                      "items": {
                          "oneOf": [
                              {
                                  "$ref": "#/components/schemas/Cyborgs.Armour"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.BodyPart"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.MeleeWeapon"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.PlayerCharacter"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.RangedWeapon"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.Components.PropValue"
                              },
                              {
                                  "$ref": "#/components/schemas/RpgContainer"
                              },
                              {
                                  "$ref": "#/components/schemas/Actions2.Action2"
                              },
                              {
                                  "$ref": "#/components/schemas/Actions2.Activity2"
                              }
                          ]
                      }
                  },
                  "contextId": {
                      "type": "string",
                      "nullable": true
                  },
                  "time": {
                      "$ref": "#/components/schemas/Time.Temporal"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.Ops.ActionStepArgs": {
              "required": [
                  "actionStep",
                  "activityId"
              ],
              "type": "object",
              "properties": {
                  "activityId": {
                      "type": "string"
                  },
                  "actionStep": {
                      "$ref": "#/components/schemas/Server.Ops.ActionStep"
                  }
              },
              "additionalProperties": false
          },
          "Server.Ops.ActionStepRun": {
              "required": [
                  "activityId",
                  "args"
              ],
              "type": "object",
              "properties": {
                  "activityId": {
                      "type": "string"
                  },
                  "args": {
                      "type": "array",
                      "items": {
                          "oneOf": [
                              {
                                  "$ref": "#/components/schemas/Reflection.Args.DefaultArg"
                              },
                              {
                                  "$ref": "#/components/schemas/Reflection.Args.DiceArg"
                              },
                              {
                                  "$ref": "#/components/schemas/Reflection.Args.IntegerArg"
                              },
                              {
                                  "$ref": "#/components/schemas/Reflection.Args.RpgObjectArg"
                              }
                          ]
                      }
                  }
              },
              "additionalProperties": false
          },
          "Server.Ops.ActivityComplete": {
              "required": [
                  "activityId"
              ],
              "type": "object",
              "properties": {
                  "activityId": {
                      "type": "string"
                  }
              },
              "additionalProperties": false
          },
          "Server.Ops.ActivityCreate": {
              "required": [
                  "actionTemplate",
                  "initiatorId",
                  "ownerId"
              ],
              "type": "object",
              "properties": {
                  "initiatorId": {
                      "type": "string"
                  },
                  "ownerId": {
                      "type": "string"
                  },
                  "actionTemplate": {
                      "type": "string"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.RpgRequest_ActionStepArgs": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Server.Ops.ActionStepArgs"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgRequest_ActionStepRun": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Server.Ops.ActionStepRun"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgRequest_ActivityComplete": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Server.Ops.ActivityComplete"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgRequest_ActivityCreate": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Server.Ops.ActivityCreate"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.RpgResponse_Action2": {
              "required": [
                  "data",
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "data": {
                      "$ref": "#/components/schemas/Actions2.Action2"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgResponse_Activity2": {
              "required": [
                  "data",
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "data": {
                      "$ref": "#/components/schemas/Actions2.Activity2"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.RpgResponse_RpgArg[]": {
              "required": [
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "data": {
                      "type": "array",
                      "items": {
                          "oneOf": [
                              {
                                  "$ref": "#/components/schemas/Reflection.Args.DefaultArg"
                              },
                              {
                                  "$ref": "#/components/schemas/Reflection.Args.DiceArg"
                              },
                              {
                                  "$ref": "#/components/schemas/Reflection.Args.IntegerArg"
                              },
                              {
                                  "$ref": "#/components/schemas/Reflection.Args.RpgObjectArg"
                              }
                          ]
                      },
                      "nullable": true
                  }
              },
              "additionalProperties": false
          },
...
      },
      "securitySchemes": {
          "Backoffice User": {
              "type": "oauth2",
              "description": "Umbraco Authentication",
              "flows": {
                  "authorizationCode": {
                      "authorizationUrl": "/umbraco/management/api/v1/security/back-office/authorize",
                      "tokenUrl": "/umbraco/management/api/v1/security/back-office/token",
                      "scopes": {}
                  }
              }
          }
      }
  }
}